requirements.txt
#
click
netifaces
textx
pyzmq
kademlia
parse
fabric
ruamel.yaml

